Refactoring
Rename dijkstra.rs to a_star_core, rename alt_router to a_star_router

Move trait definitions etc to more reasonable files. Rename traits and functions. Make some fields and methods private
